Category CRUD

+++ src/main/java/com/sc/accounting_smart_cookies/controller/CategoryController.java
@@ -64,7 +64,7 @@ public class CategoryController {
     }
 
     @PostMapping("/update/{id}")
-    public String updateCategory(@Valid @PathVariable("id") Long id, @ModelAttribute("category") CategoryDTO categoryDTO, BindingResult bindingResult, Model model){
+    public String updateCategory(@Valid @PathVariable("id") Long id, @ModelAttribute("category") CategoryDTO categoryDTO, BindingResult bindingResult){
 
         if (categoryService.hasProducts(categoryDTO)){
             bindingResult.rejectValue("description", " ", "This category already has product/products! Make sure the new description that will be provided is proper.");
@@ -78,10 +78,10 @@ public class CategoryController {
 
     @GetMapping("/delete/{id}")
     public String deleteCategory(@PathVariable("id") Long id, RedirectAttributes redirectAttributes, @ModelAttribute("category") CategoryDTO categoryDTO){
-//        if(categoryService.hasProducts(categoryDTO)){
-//            redirectAttributes.addFlashAttribute("error", "Can not be deleted! This category has product/products");
-//            return "redirect:/category-list";
-//        }
+        if(categoryService.hasProducts(categoryDTO)){
+            redirectAttributes.addFlashAttribute("error", "Can not be deleted! This category has product/products");
+            return "redirect:/categories/list";
+        }
         categoryService.deleteById(id);
         return "redirect:/categories/list";
     }
